Added an executor for the service
