+++ src/main/java/ru/mail/polis/service/ServiceFactory.java
@@ -17,7 +17,7 @@
 package ru.mail.polis.service;
 
 import ru.mail.polis.lsm.DAO;
-import ru.mail.polis.service.eldar_tim.TimService;
+import ru.mail.polis.service.eldar_tim.HttpServiceImpl;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -54,6 +54,6 @@ public final class ServiceFactory {
 
         Objects.requireNonNull(dao);
 
-        return new TimService(port, dao);
+        return new HttpServiceImpl(port, dao);
     }
 }
