--- src/main/java/ru/mail/polis/service/eldar_tim/TimService.java
@@ -1,103 +1,0 @@
-package ru.mail.polis.service.eldar_tim;
-
-import one.nio.http.HttpServer;
-import one.nio.http.HttpServerConfig;
-import one.nio.http.HttpSession;
-import one.nio.http.Param;
-import one.nio.http.Path;
-import one.nio.http.Request;
-import one.nio.http.Response;
-import one.nio.server.AcceptorConfig;
-import ru.mail.polis.lsm.DAO;
-import ru.mail.polis.lsm.Record;
-import ru.mail.polis.service.Service;
-
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
-import java.util.Iterator;
-
-/**
- * Service implementation for Stage 1 within 2021-highload-dht.
- *
- * @author Eldar Timraleev
- */
-public class TimService extends HttpServer implements Service {
-    private final DAO dao;
-
-    public TimService(final int port, final DAO dao) throws IOException {
-        super(buildHttpServerConfig(port));
-        this.dao = dao;
-    }
-
-    private static HttpServerConfig buildHttpServerConfig(final int port) {
-        final HttpServerConfig httpServerConfig = new HttpServerConfig();
-        AcceptorConfig acceptorConfig = new AcceptorConfig();
-        acceptorConfig.port = port;
-        acceptorConfig.reusePort = true;
-        acceptorConfig.deferAccept = true;
-        httpServerConfig.acceptors = new AcceptorConfig[] { acceptorConfig };
-        return httpServerConfig;
-    }
-
-    @Override
-    public void handleDefault(Request request, HttpSession session) throws IOException {
-        Response response = new Response(Response.BAD_REQUEST, Response.EMPTY);
-        session.sendResponse(response);
-    }
-
-    @Path("/v0/status")
-    public Response getStatus() {
-        return Response.ok(Response.OK);
-    }
-
-    @Path("/v0/entity")
-    public Response getEntity(
-            Request request,
-            @Param(value = "id", required = true) final String id
-    ) {
-        if (id.isBlank()) {
-            return new Response(Response.BAD_REQUEST, "Bad id".getBytes(StandardCharsets.UTF_8));
-        }
-
-        switch (request.getMethod()) {
-            case Request.METHOD_GET:
-                return get(id);
-            case Request.METHOD_PUT:
-                return put(id, request.getBody());
-            case Request.METHOD_DELETE:
-                return delete(id);
-            default:
-                return new Response(Response.METHOD_NOT_ALLOWED);
-        }
-    }
-
-    private static byte[] extractBytes(final ByteBuffer buffer) {
-        final byte[] result = new byte[buffer.remaining()];
-        buffer.get(result);
-        return result;
-    }
-
-    private Response get(String id) {
-        ByteBuffer key = ByteBuffer.wrap(id.getBytes(StandardCharsets.UTF_8));
-        final Iterator<Record> iterator = dao.range(key, DAO.nextKey(key));
-        if (iterator.hasNext()) {
-            return new Response(Response.OK, extractBytes(iterator.next().getValue()));
-        } else {
-            return new Response(Response.NOT_FOUND, Response.EMPTY);
-        }
-    }
-
-    private Response put(String id, byte[] body) {
-        ByteBuffer key = ByteBuffer.wrap(id.getBytes(StandardCharsets.UTF_8));
-        ByteBuffer value = ByteBuffer.wrap(body);
-        dao.upsert(Record.of(key, value));
-        return new Response(Response.CREATED, Response.EMPTY);
-    }
-
-    private Response delete(String id) {
-        ByteBuffer key = ByteBuffer.wrap(id.getBytes(StandardCharsets.UTF_8));
-        dao.upsert(Record.tombstone(key));
-        return new Response(Response.ACCEPTED, Response.EMPTY);
-    }
-}
